Unit tests are passing again after cleaning up the consensus module
